+++ src/main/java/http/controller/ErrorController.java
@@ -1,6 +1,6 @@
 package http.controller;
 
-import http.HttpResponse;
+import http.response.HttpResponse;
 import http.request.HttpRequest;
 
 public class ErrorController extends AbstractController {
