+++ src/main/java/http/controller/ResourcesController.java
@@ -1,6 +1,6 @@
 package http.controller;
 
-import http.HttpResponse;
+import http.response.HttpResponse;
 import http.request.HttpRequest;
 
 public class ResourcesController extends AbstractController {
