+++ src/test/java/http/HttpResponseTest.java
@@ -1,9 +1,8 @@
 package http;
 
 import http.common.HttpStatus;
+import http.response.HttpResponse;
 import org.junit.jupiter.api.Test;
-
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
